// app/src/generated/RowChoice.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { PrepareChoice } from "./PrepareChoice";
import type { PrepareRow } from "./PrepareRow";

/**
 * One row's choice.
 */
export type RowChoice = { row: PrepareRow, choice: PrepareChoice, };